Pin MCP server to released DQX + auto-sync version pins
The MCP runner now installs a pinned, published databricks-labs-dqx release
from the registry (databricks-labs-dqx[datacontract]==0.15.0) instead of
building the parent wheel from source, mirroring DQX Studio. This makes the
deployed DQX version deterministic and independent of the local working tree.

Extend the docs version-sync script (renamed update_github_urls.py ->
sync_versions.py) to also rewrite the databricks-labs-dqx==<version> pins in
app/ and mcp-server/ from __about__.py, so `make fmt` keeps them in lockstep on
release. Update the "DQX version" installation docs accordingly.

Fixes the fmt CI gate, which failed because the old source-build admonition
still contained blob/main/ GitHub URLs that the sync script rewrites.
